Effects.LoggerNS

Description

Provides namespaced logging functionality on top of MonadLogger.

Since: 0.1

Effect

class MonadLogger m => MonadLoggerNS m where

Adds namespaces to MonadLogger.

Since: 0.1

Methods

getNamespace :: HasCallStack => m Namespace

Retrieves the namespace.

Since: 0.1

localNamespace :: HasCallStack => (Namespace -> Namespace) -> m a -> m a

Locally modifies the namespace.

Since: 0.1

newtype Namespace

Logging namespace.

Since: 0.1

Constructors

MkNamespace

Fields

unNamespace :: Seq Text

Since: 0.1

addNamespace :: MonadLoggerNS m => Text -> m a -> m a

Adds to the namespace.

Since: 0.1

Levels

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

levelTrace :: LogLevel

Since: 0.1

levelFatal :: LogLevel

Since: 0.1

Logging functions

Levels

logTrace :: Q Exp

Since: 0.1

logDebug :: Q Exp

Generates a function that takes a Text and logs a LevelDebug message. Usage:

$(logDebug) "This is a debug log message"

logInfo :: Q Exp

See logDebug

logWarn :: Q Exp

See logDebug

logError :: Q Exp

See logDebug

logOther :: Text -> Q Exp

Generates a function that takes a Text and logs a LevelOther message. Usage:

$(logOther "My new level") "This is a log message"

logFatal :: Q Exp

Since: 0.1

Level checks

guardLevel

Arguments

:: Applicative f

=> LogLevel
The configured log level to check against.

-> LogLevel
The log level for this action.

-> f ()
The logging action to run if the level passes.

-> f ()

guardLevel configLvl lvl m runs m iff shouldLog configLvl lvl.

Since: 0.1

shouldLog

Arguments

:: LogLevel
The configured log level to check against.

-> LogLevel
Level for this log

-> Bool
Whether we should log

shouldLog configLvl lvl returns true iff configLvl <= lvl. Uses LogLevel's built-in ordering with special cases for Trace (LevelOther Trace) and Fatal (LevelOther Fatal). The ad-hoc ordering is thus:

  LevelOther "Trace"
    < LevelDebug
    < LevelInfo
    < LevelWarn
    < LevelError
    < LevelOther "Fatal"
    < LevelOther "<any>"

In other words, LogLevel's usual Ord is respected, with the additional cases. Note that any other LevelOther "custom" sit at the the highest level and compare via Text's Ord, just like LogLevel's usual Ord.

Since: 0.1

Formatting

data LogFormatter

Formatter for logs.

Since: 0.1

Constructors

MkLogFormatter

Fields

newline :: !Bool

If true, append a newline.

Since: 0.1

locStrategy :: !LocStrategy

How to log the code location.

Since: 0.1

timezone :: !Bool

Whether to include the timezone in the timestamp.

Since: 0.1

defaultLogFormatter :: Loc -> LogFormatter

LogFormatter with:

newline = True
locStrategy = LocPartial loc
timezone = False

Since: 0.1

data LocStrategy

Determines how we log location data.

Since: 0.1

Constructors

LocPartial !Loc
Logs the location with filename, line, col.

Since: 0.1

LocStable !Loc
Logs the location with filename.

Since: 0.1

LocNone
No location logging.

Since: 0.1

formatLog :: (HasCallStack, MonadLoggerNS m, MonadTime m, ToLogStr msg) => LogFormatter -> LogLevel -> msg -> m LogStr

Produces a formatted LogStr.

Since: 0.1

LogStr

logStrToBs :: LogStr -> ByteString

Since: 0.1

logStrToText :: LogStr -> Text

Since: 0.1

Optics

LogLevels

_LevelTrace :: Prism' LogLevel ()

Since: 0.1

_LevelInfo :: Prism' LogLevel ()

Since: 0.1

_LevelDebug :: Prism' LogLevel ()

Since: 0.1

_LevelWarn :: Prism' LogLevel ()

Since: 0.1

_LevelError :: Prism' LogLevel ()

Since: 0.1

_LevelOther :: Prism' LogLevel Text

Since: 0.1

_LevelFatal :: Prism' LogLevel ()

Since: 0.1

LocStrategy

_LocPartial :: Prism' LocStrategy Loc

Since: 0.1

_LocStable :: Prism' LocStrategy Loc

Since: 0.1

_LocNone :: Prism' LocStrategy ()

Since: 0.1

Reexports

class Monad m => MonadLogger (m :: Type -> Type) where

A Monad which has the ability to log messages in some manner.

Minimal complete definition

Nothing

Methods

monadLoggerLog :: ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> m ()

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

data Loc